+++ app/uk/gov/hmrc/residencenilratebandcalculator/microserviceWiring.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.inheritancetaxresidencenilratebandcalculator
+package uk.gov.hmrc.residencenilratebandcalculator
 
 import uk.gov.hmrc.play.audit.http.config.LoadAuditingConfig
 import uk.gov.hmrc.play.audit.http.connector.AuditConnector
