+++ app/uk/gov/hmrc/residencenilratebandcalculator/models/Percent.scala
@@ -14,10 +14,15 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.inheritancetaxresidencenilratebandcalculator.models
+package uk.gov.hmrc.residencenilratebandcalculator.models
 
-import org.joda.time.LocalDate
+case class Percent(value: Double) extends Ordered[Percent] {
 
-object TaperBand extends Band {
-  def apply(date: LocalDate): Int = 2000000
+  def asDecimal: Double = value / 100
+
+  override def compare(that: Percent) = (this.value - that.value) match {
+    case x if x > 0 => 1
+    case x if x == 0 => 0
+    case _ => -1
+  }
 }
