+++ app/uk/gov/hmrc/residencenilratebandcalculator/models/ResidenceNilRateBand.scala
@@ -14,6 +14,19 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.inheritancetaxresidencenilratebandcalculator.models
+package uk.gov.hmrc.residencenilratebandcalculator.models
 
-case class CalculationResult (residenceNilRateAmount: Int, carryForwardAmount: Int)
+import org.joda.time.LocalDate
+
+object ResidenceNilRateBand extends Band {
+  def apply(date: LocalDate): Int = {
+    val bands = Map(
+      new LocalDate(2018, 4, 5) -> 125000,
+      new LocalDate(2019, 4, 5) -> 150000,
+      new LocalDate(2020, 4, 5) -> 175000)
+
+    val bandDate = getHighestDateBefore(date, bands)
+
+    bandDate.fold(100000) { d => bands(d) }
+  }
+}
