+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.inheritancetaxresidencenilratebandcalculator.controllers.MicroserviceHelloWorld.hello
+GET        /hello-world             uk.gov.hmrc.residencenilratebandcalculator.controllers.MicroserviceHelloWorld.hello
