+++ conf/application.conf
@@ -15,7 +15,7 @@
 # This is the main configuration file for the application.
 # ~~~~~
 
-appName=inheritance-tax-residence-nil-rate-band-calculator
+appName=residence-nil-rate-band-calculator
 
 # Play Modules
 # ~~~~
@@ -54,7 +54,7 @@ application.langs="en"
 # ~~~~~
 # Define the Global object class for this application.
 # Default to Global in the root package.
-application.global=uk.gov.hmrc.inheritancetaxresidencenilratebandcalculator.MicroserviceGlobal
+application.global=uk.gov.hmrc.residencenilratebandcalculator.MicroserviceGlobal
 
 # Router
 # ~~~~~
@@ -85,7 +85,7 @@ controllers {
       needsAuditing = false
     }
 
-  uk.gov.hmrc.inheritancetaxresidencenilratebandcalculator.controllers.MicroserviceHelloWorld = {
+  uk.gov.hmrc.residencenilratebandcalculator.controllers.MicroserviceHelloWorld = {
       needsAuth = false
       needsLogging = false
       needsAuditing = false
