+++ conf/prod.routes
@@ -1,5 +1,5 @@
 # Add all the application routes to the app.routes file
-->         /inheritance-tax-residence-nil-rate-band-calculator              app.Routes
+->         /residence-nil-rate-band-calculator              app.Routes
 ->         /                          health.Routes
 
 GET        /admin/metrics             @com.kenshoo.play.metrics.MetricsController.metrics
